feat: Integrate ZSpotifyGUI into the main application
This commit integrates the outdated `ZSpotifyGUI` into the modern
`zotify` command-line application, creating a single, unified program
with a functional graphical user interface.

Key Changes:
- Restructures the project by moving all GUI code into a `zotify/gui`
  module.
- Replaces the broken username/password login with the modern OAuth flow
  from the core backend.
- Connects the GUI's search and download functionality to the backend API.
- Implements real-time download progress reporting for tracks, albums,
  and playlists.
- Adds a functional settings dialog to manage configuration.
- Implements the "Your Library" section to show liked and local songs.
- Adds PyInstaller support for packaging the GUI as a standalone
  executable.

+++ zotify/gui/worker.py
@@ -2,7 +2,6 @@ import sys
 import traceback
 import logging
 from PyQt5.QtCore import QObject, pyqtSignal, QRunnable
-from const import TRACKS, ARTISTS, ALBUMS, PLAYLISTS
 
 logger = logging.getLogger(__name__)
 
@@ -13,7 +12,7 @@ class WorkerSignals(QObject):
     result = pyqtSignal(object)
 
 class MusicSignals(WorkerSignals):
-    update = pyqtSignal(float,int,int)
+    update = pyqtSignal(int, int, int)
 
 class Worker(QRunnable):
     """
